+++ integration-test/src/test/resources/generate-model/src/main/kotlin/target/test/generatemodel/TestModel.kt
@@ -1,0 +1,13 @@
+package target.test.generatemodel
+
+import target.annotation.Validatable
+
+@Validatable
+data class TestModel(
+    val id: PositiveInt,
+    val isEnabled: Boolean,
+    val updated: Long,
+    val created: Long,
+) {
+    companion object
+}