+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.8.0
+
+Feature:
+
+- Added initial `integration-test` package.
+- Updated maven central publishing.
+- Kotlin 2.2.20
+- KSP 2.2.20-2.0.2
+
 ## 0.7.0
 
 Feature: